Dropping rejected for second MAF; Filtering goes to sep directory; No bam required for anno and qc

+++ source/bcbio_runner.py
@@ -8,7 +8,7 @@ from source.file_utils import verify_dir, verify_file, file_transaction, make_tm
 from source.tools_from_cnf import get_tool_cmdline
 
 from source.utils_from_bcbio import file_exists, safe_mkdir, add_suffix
-from source.logger import info
+from source.logger import info, err
 from source.ngscat.bed_file import verify_bam
 
 
@@ -44,7 +44,7 @@ class Steps(list):
 
     def step(self, name, script=None, interpreter='python', param_line=None):
         if name not in self:
-            return None
+            return Step(name, None)
 
         cmd = get_tool_cmdline(self.cnf, interpreter, script or name)
         if not cmd:
@@ -65,6 +65,7 @@ class Runner():
 
         self.varannotate = None
         self.varqc = None
+        self.varqc_after = None
         self.varfilter = None
         self.targetcov = None
         self.ngscat = None
@@ -94,15 +95,19 @@ class Runner():
         self.varannotate = self.steps.step(
             name='VarAnnotate',
             script='varannotate.py',
-            param_line=spec_params + ' --vcf \'{vcf}\' --bam \'{bam}\' -o \'{output_dir}\' -s \'{sample}\' --work-dir \'' + join(cnf.work_dir, 'varannotate') + '\'')
+            param_line=spec_params + ' --vcf \'{vcf}\' {bam_cmdline} -o \'{output_dir}\' -s \'{sample}\' --work-dir \'' + join(cnf.work_dir, 'varannotate') + '\'')
         self.varqc = self.steps.step(
             name='VarQC',
             script='varqc.py',
             param_line=spec_params + ' --vcf \'{vcf}\' -o \'{output_dir}\' -s \'{sample}\' --work-dir \'' + join(cnf.work_dir, 'varqc') + '\'')
+        self.varqc_after = self.steps.step(
+            name='VarQC_after',
+            script='varqc.py',
+            param_line=spec_params + ' --vcf \'{vcf}\' -o \'{output_dir}\' -s \'{sample}\' --work-dir \'' + join(cnf.work_dir, 'varqc_after') + '\'')
         self.varfilter = self.steps.step(
             name='VarFilter',
             script='varfilter.py',
-            param_line=spec_params + ' --vcf \'{vcf}\' -o \'{output_dir}\' -s \'{sample}\' --clean --work-dir \'' + join(cnf.work_dir, 'varfilter') + '\'')
+            param_line=spec_params + ' --vcf \'{vcf}\' -o \'{output_dir}\' -s \'{sample}\' --work-dir \'' + join(cnf.work_dir, 'varfilter') + '\'')
         self.targetcov = self.steps.step(
             name='TargetCov',
             script='targetcov.py',
@@ -125,22 +130,18 @@ class Runner():
             self.varqc_summary = self.steps.step(
                 name='VarQC_summary',
                 script='varqc_summary.py',
-                param_line=cnfs_line + ' -o \'{output_dir}\' -d \'' + self.dir + '\' -s \'{samples}\' -n ' +
-                           self.varqc.name.lower() + ' --vcf-suf ' + ','.join(all_suffixes) +
-                           ' --work-dir \'' + join(cnf.work_dir, 'varqc_summary') + '\'')
+                param_line=cnfs_line + ' -o \'{output_dir}\' -d \'' + self.dir + '\' -s \'{samples}\''
+                           ' -n varqc --vcf-suf ' + ','.join(all_suffixes) + ' --work-dir \'' + join(cnf.work_dir, 'varqc_summary') + '\'')
         if self.targetcov:
             self.steps.append('TargetCov_summary')
             self.targetcov_summary = self.steps.step(
                 name='TargetCov_summary',
                 script='targetcov_summary.py',
-                param_line=cnfs_line + ' -o \'{output_dir}\' -d \'' + self.dir +
-                           '\' -s \'{samples}\' -n ' + self.targetcov.name.lower() +
-                           ' --work-dir \'' + join(cnf.work_dir, 'targetcov_summary') + '\'')
+                param_line=cnfs_line + ' -o \'{output_dir}\' -d \'' + self.dir + '\' -s \'{samples}\''
+                           ' -n targetcov --work-dir \'' + join(cnf.work_dir, 'targetcov_summary') + '\'')
 
-    def submit(self, step, sample_name='', suf=None, create_dir=False, out_fpath=None,
+    def submit_if_needed(self, step, sample_name='', suf=None, create_dir=True, out_fpath=None,
                wait_for_steps=list(), **kwargs):
-        if not step or step.name not in self.steps:
-            return None
 
         output_dirpath = self.dir
         if sample_name:
@@ -154,15 +155,28 @@ class Runner():
 
         if create_dir:
             output_dirpath = join(output_dirpath, step.name.lower())
-            safe_mkdir(output_dirpath)
             log_fpath = join(output_dirpath, 'log' + ('_' + suf if suf else ''))
+
+        # Skipping steps
+        if not step or step.name not in self.steps:
+            return output_dirpath
+
+        if create_dir:
+            safe_mkdir(output_dirpath)
 
         out_fpath = out_fpath or log_fpath
 
         if isfile(out_fpath):
-            os.remove(out_fpath)
+            try:
+                os.remove(out_fpath)
+            except OSError:
+                err('Cannot remove log file ' + out_fpath + ', probably permission denied.')
+
         if isfile(log_fpath):
-            os.remove(log_fpath)
+            try:
+                os.remove(log_fpath)
+            except OSError:
+                err('Cannot remove log file ' + out_fpath + ', probably permission denied.')
 
         hold_jid_line = '-hold_jid ' + ','.join(wait_for_steps or ['_'])
 
@@ -245,8 +259,12 @@ class Runner():
                 sys.exit(1)
 
             bam_fpath = join(sample_dirpath, sample + '-ready.bam')
-            if not verify_bam(bam_fpath):
-                sys.exit(1)
+            if 'TargetCov' in self.steps or 'NGScat' in self.steps or 'Qualimap' in self.steps:
+                if not verify_bam(bam_fpath):
+                    sys.exit(1)
+            else:
+                if not file_exists(bam_fpath):
+                    bam_fpath = None
 
             for variant_caller in sample_info['algorithm'].get('variantcaller') or []:
                 vcf_fpath = join(sample_dirpath, sample + '-' + variant_caller + '.vcf')
@@ -267,11 +285,11 @@ class Runner():
 
                 self._process_vcf(sample, sample_dirpath, bam_fpath, vcf_fpath, variant_caller)
 
-            self.submit(self.targetcov, sample, create_dir=True, bam=bam_fpath, bed=bed_fpath, sample=sample)
+            self.submit_if_needed(self.targetcov, sample, bam=bam_fpath, bed=bed_fpath, sample=sample)
 
-            self.submit(self.ngscat, sample, create_dir=True, bam=bam_fpath, bed=bed_fpath, sample=sample)
+            self.submit_if_needed(self.ngscat, sample, bam=bam_fpath, bed=bed_fpath, sample=sample)
 
-            self.submit(self.qualimap, sample, create_dir=True, bam=bam_fpath, bed=qualimap_bed_fpath, sample=sample)
+            self.submit_if_needed(self.qualimap, sample, bam=bam_fpath, bed=qualimap_bed_fpath, sample=sample)
 
             if self.cnf.verbose:
                 info('-' * 70)
@@ -293,22 +311,20 @@ class Runner():
                 sample = sample_info['description']
                 f.write(sample + '\n')
 
-        if self.varqc:
-            self.submit(
-                self.varqc_summary,
-                create_dir=True,
-                wait_for_steps=[self.varqc.job_name(d['description'], v)
-                                for d in self.bcbio_cnf.details
-                                for v in all_variantcallers],
-                samples=samples_fpath)
+        self.submit_if_needed(
+            self.varqc_summary,
+            wait_for_steps=[
+                self.varqc.job_name(d['description'], v)
+                for d in self.bcbio_cnf.details
+                for v in all_variantcallers],
+            samples=samples_fpath)
 
-        if self.targetcov_summary:
-            self.submit(
-                self.targetcov_summary,
-                create_dir=True,
-                wait_for_steps=[self.targetcov.job_name(d['description'])
-                                for d in self.bcbio_cnf.details],
-                samples=samples_fpath)
+        self.submit_if_needed(
+            self.targetcov_summary,
+            wait_for_steps=[
+                self.targetcov.job_name(d['description'])
+                for d in self.bcbio_cnf.details],
+            samples=samples_fpath)
 
         if not self.cnf.verbose:
             print ''
@@ -316,22 +332,20 @@ class Runner():
             info('Done.')
 
     def _process_vcf(self, sample, sample_dirpath, bam_fpath, vcf_fpath, suf):
-        if self.varqc:
-            self.submit(
-                self.varqc, sample, suf=suf, create_dir=True,
-                vcf=vcf_fpath, sample=sample + '-' + suf)
+        self.submit_if_needed(self.varqc, sample, suf=suf, vcf=vcf_fpath, sample=sample + '-' + suf)
 
-        if self.varannotate:
-            anno_dirpath = self.submit(
-                self.varannotate, sample, suf=suf, create_dir=True,
-                vcf=vcf_fpath, bam=bam_fpath, sample=sample + '-' + suf)
-            annotated_vcf_fpath = join(anno_dirpath, basename(add_suffix(vcf_fpath, 'anno')))
+        bam_cmdline = '--bam ' + bam_fpath if bam_fpath else ''
+        anno_dirpath = self.submit_if_needed(self.varannotate, sample, suf=suf, vcf=vcf_fpath, bam_cmdline=bam_cmdline, sample=sample + '-' + suf)
 
-            if self.varfilter:
-                filtered_vcf_fpath = join(sample_dirpath,
-                    basename(add_suffix(annotated_vcf_fpath, 'filt')))
-                if file_exists(filtered_vcf_fpath):
-                    os.remove(filtered_vcf_fpath)
-                self.submit(self.varfilter, sample, suf=suf, create_dir=False,
-                    wait_for_steps=[self.varannotate.job_name(sample, suf)],
-                    vcf=annotated_vcf_fpath, sample=sample + '-' + suf)
+        annotated_vcf_fpath = join(anno_dirpath, basename(add_suffix(vcf_fpath, 'anno')))
+
+        filter_dirpath = self.submit_if_needed(self.varfilter, sample, suf=suf,
+            wait_for_steps=[self.varannotate.job_name(sample, suf)],
+            vcf=annotated_vcf_fpath, sample=sample + '-' + suf)
+
+        filtered_vcf_fpath = join(filter_dirpath, basename(add_suffix(annotated_vcf_fpath, 'filt')))
+
+        self.submit_if_needed(
+            self.varqc_after, sample, suf=suf,
+            wait_for_steps=[self.varfilter.job_name(sample, suf)] if self.varfilter.name in self.steps else [],
+            vcf=filtered_vcf_fpath, sample=sample + '-' + suf)
